fix entity manager
