+++ homeassistant/components/geo_json_events/__init__.py
@@ -1,12 +1,7 @@
 """The GeoJSON events component."""
 from __future__ import annotations
 
-from collections.abc import Callable
-from datetime import datetime, timedelta
 import logging
-
-from aio_geojson_generic_client import GenericFeedManager
-from aio_geojson_generic_client.feed_entry import GenericFeedEntry
 
 from homeassistant.config_entries import SOURCE_IMPORT, ConfigEntry
 from homeassistant.const import (
@@ -17,24 +12,15 @@ from homeassistant.const import (
     CONF_URL,
     Platform,
 )
-from homeassistant.core import HomeAssistant, callback
-from homeassistant.helpers import aiohttp_client
-from homeassistant.helpers.dispatcher import async_dispatcher_send
+from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_registry import (
     async_entries_for_config_entry,
     async_get,
 )
-from homeassistant.helpers.event import async_track_time_interval
 from homeassistant.helpers.typing import ConfigType
 
-from .const import (
-    DEFAULT_SCAN_INTERVAL,
-    DOMAIN,
-    FEED,
-    PLATFORMS,
-    SIGNAL_DELETE_ENTITY,
-    SIGNAL_UPDATE_ENTITY,
-)
+from .const import DEFAULT_SCAN_INTERVAL, DOMAIN, FEED, PLATFORMS
+from .manager import GeoJsonFeedEntityManager
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -72,7 +58,7 @@ async def async_setup_entry(hass: HomeAssistant, config_entry: ConfigEntry) -> b
     feeds = hass.data[DOMAIN].setdefault(FEED, {})
 
     # Create feed entity manager for all platforms.
-    manager = GeoJsonEventsFeedEntityManager(hass, config_entry)
+    manager = GeoJsonFeedEntityManager(hass, config_entry)
     feeds[config_entry.entry_id] = manager
     _LOGGER.debug("Feed entity manager added for %s", config_entry.entry_id)
     # Remove orphaned geo_location entities.
@@ -96,95 +82,3 @@ async def async_unload_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
         manager = hass.data[DOMAIN][FEED].pop(entry.entry_id)
         await manager.async_stop()
     return unload_ok
-
-
-class GeoJsonEventsFeedEntityManager:
-    """Feed Entity Manager for GeoJSON events feed."""
-
-    def __init__(
-        self,
-        hass: HomeAssistant,
-        config_entry: ConfigEntry,
-    ) -> None:
-        """Initialize the Feed Entity Manager."""
-        self._hass = hass
-        self._config_entry = config_entry
-        coordinates: tuple[float, float] = (
-            config_entry.data.get(CONF_LATITUDE, hass.config.latitude),
-            config_entry.data.get(CONF_LONGITUDE, hass.config.longitude),
-        )
-        self._url = config_entry.data[CONF_URL]
-        radius_in_km = config_entry.data[CONF_RADIUS]
-        websession = aiohttp_client.async_get_clientsession(hass)
-        self._feed_manager = GenericFeedManager(
-            websession,
-            self._generate_entity,
-            self._update_entity,
-            self._remove_entity,
-            coordinates,
-            self._url,
-            filter_radius=radius_in_km,
-        )
-        self._config_entry_id = config_entry.entry_id
-        self._scan_interval = timedelta(seconds=DEFAULT_SCAN_INTERVAL)
-        self._track_time_remove_callback: Callable[[], None] | None = None
-        self.listeners: list[Callable[[], None]] = []
-
-    async def async_init(self) -> None:
-        """Schedule initial and regular updates based on configured time interval."""
-
-        await self._hass.config_entries.async_forward_entry_setups(
-            self._config_entry, PLATFORMS
-        )
-
-        async def update(event_time: datetime) -> None:
-            """Update."""
-            await self.async_update()
-
-        # Trigger updates at regular intervals.
-        self._track_time_remove_callback = async_track_time_interval(
-            self._hass, update, self._scan_interval
-        )
-
-        _LOGGER.debug("Feed entity manager initialized")
-
-    async def async_update(self) -> None:
-        """Refresh data."""
-        await self._feed_manager.update()
-        _LOGGER.debug("Feed entity manager updated")
-
-    async def async_stop(self) -> None:
-        """Stop this feed entity manager from refreshing."""
-        for unsub_dispatcher in self.listeners:
-            unsub_dispatcher()
-        self.listeners = []
-        if self._track_time_remove_callback:
-            self._track_time_remove_callback()
-        _LOGGER.debug("Feed entity manager stopped")
-
-    @callback
-    def async_event_new_entity(self):
-        """Return manager specific event to signal new entity."""
-        return f"{DOMAIN}_new_geolocation_{self._config_entry_id}"
-
-    def get_entry(self, external_id: str) -> GenericFeedEntry | None:
-        """Get feed entry by external id."""
-        return self._feed_manager.feed_entries.get(external_id)
-
-    async def _generate_entity(self, external_id: str) -> None:
-        """Generate new entity."""
-        async_dispatcher_send(
-            self._hass,
-            self.async_event_new_entity(),
-            self,
-            self._config_entry.unique_id,
-            external_id,
-        )
-
-    async def _update_entity(self, external_id: str) -> None:
-        """Update entity."""
-        async_dispatcher_send(self._hass, SIGNAL_UPDATE_ENTITY.format(external_id))
-
-    async def _remove_entity(self, external_id: str) -> None:
-        """Remove entity."""
-        async_dispatcher_send(self._hass, SIGNAL_DELETE_ENTITY.format(external_id))
